Fix odr_cal_reg_set/get API, change type uint8 to int8

+++ asm330lhh_reg.c
@@ -1235,7 +1235,7 @@ int32_t asm330lhh_device_conf_get(const stmdev_ctx_t *ctx, uint8_t *val)
   * @retval        Interface status (MANDATORY: return 0 -> no Error).
   *
   */
-int32_t asm330lhh_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
+int32_t asm330lhh_odr_cal_reg_set(const stmdev_ctx_t *ctx, int8_t val)
 {
   asm330lhh_internal_freq_fine_t internal_freq_fine;
   int32_t ret;
@@ -1244,7 +1244,7 @@ int32_t asm330lhh_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
                            (uint8_t *)&internal_freq_fine, 1);
   if (ret == 0)
   {
-    internal_freq_fine.freq_fine = (uint8_t)val;
+    internal_freq_fine.freq_fine = val;
     ret = asm330lhh_write_reg(ctx, ASM330LHH_INTERNAL_FREQ_FINE,
                               (uint8_t *)&internal_freq_fine, 1);
   }
@@ -1261,7 +1261,7 @@ int32_t asm330lhh_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
   * @retval        Interface status (MANDATORY: return 0 -> no Error).
   *
   */
-int32_t asm330lhh_odr_cal_reg_get(const stmdev_ctx_t *ctx, uint8_t *val)
+int32_t asm330lhh_odr_cal_reg_get(const stmdev_ctx_t *ctx, int8_t *val)
 {
   asm330lhh_internal_freq_fine_t internal_freq_fine;
   int32_t ret;
@@ -1273,7 +1273,7 @@ int32_t asm330lhh_odr_cal_reg_get(const stmdev_ctx_t *ctx, uint8_t *val)
     return ret;
   }
 
-  *val = internal_freq_fine.freq_fine;
+  *val = (int8_t)internal_freq_fine.freq_fine;
 
   return ret;
 }
